#174 feat: move view extensions
