+++ StepSquad/StepSquad/View/ShowNewBirdView.swift
@@ -44,39 +44,6 @@ struct ShowNewBirdView: View {
     ShowNewBirdView()
 }
 
-extension View {
-    // This function changes our View to UIView, then calls another function
-    // to convert the newly-made UIView to a UIImage.
-    public func asUIImage() -> UIImage {
-        let controller = UIHostingController(rootView: self)
-        
-        // Set the background to be transparent incase the image is a PNG, WebP or (Static) GIF
-        controller.view.backgroundColor = .clear
-        
-        controller.view.frame = CGRect(x: 0, y: CGFloat(Int.max), width: 1, height: 1)
-        UIApplication.shared.windows.first!.rootViewController?.view.addSubview(controller.view)
-        
-        let size = controller.sizeThatFits(in: UIScreen.main.bounds.size)
-        controller.view.bounds = CGRect(origin: .zero, size: size)
-        controller.view.sizeToFit()
-        
-        // here is the call to the function that converts UIView to UIImage: `.asUIImage()`
-        let image = controller.view.asUIImage()
-        controller.view.removeFromSuperview()
-        return image
-    }
-}
-
-extension UIView {
-    // This is the function to convert UIView to UIImage
-    public func asUIImage() -> UIImage {
-        let renderer = UIGraphicsImageRenderer(bounds: bounds)
-        return renderer.image { rendererContext in
-            layer.render(in: rendererContext.cgContext)
-        }
-    }
-}
-
 struct ItemsConfettiView: View {
     var body: some View {
         ZStack {
@@ -136,6 +103,7 @@ struct EmptyNestView: View {
                     .font(.title)
                     .fontWeight(.bold)
                     .foregroundStyle(Color(hex: 0x638D48))
+                    .animation(.linear(duration: 1).delay(1))
             }
             RiveViewModel(fileName: "reset").view()
         }
